Update 'correct' command
1. correct_reads.c: fix bug printing wrong opening bkmer
2. correct_reads.c: print read bases (in lowercase) instead of N
3. ctx_correct.c take gap_variance, gap_wiggle (-d,-D) as cmd arguments
4. libs/cJSON: store numbers as `long` instead of `int`
5. db_graph.c: add db_graph_next_nodes_in_col() useful function
6. read_thread_cmd.c: only use max_context 1 if no paths loaded

Other:
* Fixed some warnings from CoverityScan

+++ libs/cJSON/cJSON.h
@@ -47,7 +47,7 @@ typedef struct cJSON {
 	int type;					/* The type of the item, as above. */
 
 	char *valuestring;			/* The item's string, if type==cJSON_String */
-	int valueint;				/* The item's number, if type==cJSON_Number */
+	long valueint;				/* The item's number, if type==cJSON_Number */
 	double valuedouble;			/* The item's number, if type==cJSON_Number */
 
 	char *string;				/* The item's name string, if this item is the child of, or is in the list of subitems of an object. */
